+++ data/scripts/entity/init.lua
@@ -8,5 +8,9 @@ if onServer() then
 
     if not entity.aiOwned and (entity.isShip or entity.isStation or entity.isDrone) and (entity.playerOwned or entity.allianceOwned) then
         entity:addScriptOnce("data/scripts/entity/TrashMan.lua")
+        entity:addScriptOnce("data/scripts/entity/captainelitetraits.lua")
+        if entity.isStation then
+            entity:addScriptOnce("data/scripts/entity/stationgovernor.lua")
+        end
     end
 end